+++ src/main/java/com/hnm/billing/service/BillService.java
@@ -16,6 +16,8 @@ public interface BillService {
     Connection saveConnection(Connection connection, long supplierId);
     List<Connection> getConnectionsByUserId(long userId);
     List<Bill> getBillsByUserIdAndConnectionType(long userId, String connectionType);
-    String payBill(long billId, long lateFee);
+
+    String payBill(long billId, Double lateFee);
+
     List<BillDTO> getAllBills(String connectionType, String billStatus);
 }
